ResourseNotFoundException, DatabaseException

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -4,10 +4,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourseNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -21,7 +26,7 @@ public class ProductService {
 
     @Transactional (readOnly = true)
     public ProductDTO findById(Long id){
-        ProductDTO dto = new ProductDTO(repository.findById(id).get());
+        ProductDTO dto = new ProductDTO(repository.findById(id).orElseThrow(() -> new ResourseNotFoundException("Recurso nao encontrado")));
         return dto;
     }
 
@@ -42,11 +47,17 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto,entity);
-        entity = repository.save(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto,entity);
+            entity = repository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+
+        } catch (EntityNotFoundException e){
+            throw new ResourseNotFoundException("Recurso nao encontrado");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -56,9 +67,17 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourseNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
 }
